remove some msvc-warnings

+++ lib/SmallChange/nodekits/SmDynamicObjectKit.cpp
@@ -117,7 +117,7 @@ SmDynamicObjectKit::SmDynamicObjectKit(void)
   
   this->setGeometryVisibility(TRUE);
   SoRotation * rn = new SoRotation;
-  rn->rotation.setValue(SbVec3f(1,0,0), M_PI/2.0);
+  rn->rotation.setValue(SbVec3f(1,0,0), float(M_PI/2.0));
   this->setAnyPart("stdRotation", rn);
 }
 
@@ -185,7 +185,7 @@ SmDynamicObjectKit::reset(void)
   this->setAnyPart("file", NULL, TRUE);
   this->setAnyPart("childList", NULL, TRUE);
   SoRotation * rn = new SoRotation;
-  rn->rotation.setValue(SbVec3f(1,0,0), M_PI/2.0);
+  rn->rotation.setValue(SbVec3f(1,0,0), float(M_PI/2.0));
   this->setAnyPart("stdRotation", rn);
   this->setGeometryVisibility(TRUE);
   
@@ -227,7 +227,7 @@ SmDynamicObjectKit::updateScene(void)
   SbVec3d thispos = this->position.getValue();
   if (dok_elevation_cb) {
     if (this->hasRelativeElevation.getValue()) {
-      float elevation = thispos[2];
+      float elevation = float(thispos[2]);
       float terrain_elev;
       if (dok_elevation_cb(thispos[0], thispos[1], terrain_elev)) {
         thispos[2] = elevation + terrain_elev;
@@ -244,9 +244,9 @@ SmDynamicObjectKit::updateScene(void)
     SoTranslation * pos = (SoTranslation *)this->getAnyPart("relativePosition", TRUE);
     assert(pos);
     SbVec3f posvec;
-    posvec[0] = thispos[0];
-    posvec[1] = thispos[1];
-    posvec[2] = thispos[2];
+    posvec[0] = float(thispos[0]);
+    posvec[1] = float(thispos[1]);
+    posvec[2] = float(thispos[2]);
     pos->enableNotify(FALSE);
     pos->translation.setValue(posvec);
     pos->enableNotify(TRUE);
@@ -263,9 +263,9 @@ SmDynamicObjectKit::updateScene(void)
     pos->enableNotify(TRUE);
   }
   // Set rotation from heading, pitch, roll fields
-  SbRotation h(SbVec3f(0,0,1), -1 * (M_PI * this->heading.getValue() / 180.0));
-  SbRotation p(SbVec3f(1,0,0), (M_PI * this->pitch.getValue() / 180.0));
-  SbRotation r(SbVec3f(0,1,0), (M_PI * this->roll.getValue() / 180.0));
+  SbRotation h(SbVec3f(0,0,1), float(-1 * (M_PI * this->heading.getValue() / 180.0)));
+  SbRotation p(SbVec3f(1,0,0), float(M_PI * this->pitch.getValue() / 180.0));
+  SbRotation r(SbVec3f(0,1,0), float(M_PI * this->roll.getValue() / 180.0));
   SbRotation newrot = r*p*h;
   SoRotation * rot = (SoRotation *)this->getAnyPart("rotation", TRUE);
   assert(rot);
